+++ src/cdf/Map.java
@@ -112,7 +112,7 @@ public final class Map extends Variable {
         // valeur x
         sb.append(getShortName().substring(0, getShortName().length() - 2) + "X_A" + " = ");
         sb.append("[");
-        for (int x = 1; x < dimX; x++) {
+        for (short x = 1; x < dimX; x++) {
             if (x > 1) {
                 sb.append(" " + getValue(0, x));
             } else {
@@ -127,7 +127,7 @@ public final class Map extends Variable {
         sb.append("\n");
         sb.append(getShortName().substring(0, getShortName().length() - 2) + "Y_A" + " = ");
         sb.append("[");
-        for (int x = 1; x < dimY; x++) {
+        for (short x = 1; x < dimY; x++) {
             if (x > 1) {
                 sb.append(" " + getValue(x, 0));
             } else {
@@ -145,8 +145,8 @@ public final class Map extends Variable {
             sb.append("\n");
             sb.append(getShortName() + " = ");
             sb.append("[");
-            for (int x = 1; x < dimX; x++) {
-                for (int y = 1; y < dimY; y++) {
+            for (short x = 1; x < dimX; x++) {
+                for (short y = 1; y < dimY; y++) {
                     if (y > 1) {
                         sb.append(" " + getValue(y, x));
                     } else {
@@ -167,8 +167,8 @@ public final class Map extends Variable {
             sb.append("\n");
             sb.append(getShortName() + " = ");
             sb.append("[");
-            for (int y = 1; y < dimY; y++) {
-                for (int x = 1; x < dimX; x++) {
+            for (short y = 1; y < dimY; y++) {
+                for (short x = 1; x < dimX; x++) {
                     if (x > 1) {
                         sb.append(" " + getValue(y, x));
                     } else {
